+++ db/migrations/3_alter_user.up.sql
@@ -1,0 +1,5 @@
+BEGIN;
+
+INSERT INTO "public"."users"  (id, username, password, max_task_per_day) VALUES('firstUser','manabie' ,'example', 5);
+
+COMMIT;